+++ graphql/codegen/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-codegen@4.1.3...@constructive-io/graphql-codegen@4.2.0) (2026-02-23)
+
+### Features
+
+- **codegen:** add --select flag to custom CLI commands using nested-obj ([e4095ee](https://github.com/constructive-io/constructive/commit/e4095ee9af14900a16bda35cd482e7ff86c46e95))
+
+## [4.1.3](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-codegen@4.1.2...@constructive-io/graphql-codegen@4.1.3) (2026-02-23)
+
+### Bug Fixes
+
+- **codegen:** add select param and JSON input parsing to custom command generator ([58e631e](https://github.com/constructive-io/constructive/commit/58e631ef9b672ce48daa8a689252602a3c3ee5fd))
+
+## [4.1.2](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-codegen@4.1.1...@constructive-io/graphql-codegen@4.1.2) (2026-02-23)
+
+### Bug Fixes
+
+- **codegen:** fix CLI codegen bugs - ORM signatures, String.call, type coercion, field filtering ([91d58ad](https://github.com/constructive-io/constructive/commit/91d58adf34dd10dc55ee1068787b78f922234aa9))
+
 ## [4.1.1](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-codegen@4.1.0...@constructive-io/graphql-codegen@4.1.1) (2026-02-19)
 
 **Note:** Version bump only for package @constructive-io/graphql-codegen