+++ graphql/test-app/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.1.3...@constructive-io/test-codegen-app@1.1.4) (2026-02-23)
+
+**Note:** Version bump only for package @constructive-io/test-codegen-app
+
+## [1.1.3](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.1.2...@constructive-io/test-codegen-app@1.1.3) (2026-02-23)
+
+**Note:** Version bump only for package @constructive-io/test-codegen-app
+
+## [1.1.2](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.1.1...@constructive-io/test-codegen-app@1.1.2) (2026-02-23)
+
+**Note:** Version bump only for package @constructive-io/test-codegen-app
+
 ## [1.1.1](https://github.com/constructive-io/constructive/compare/@constructive-io/test-codegen-app@1.1.0...@constructive-io/test-codegen-app@1.1.1) (2026-02-19)
 
 **Note:** Version bump only for package @constructive-io/test-codegen-app